Adjusted file structure
