+++ frontend/src/liquity/charts/StakingLUSD.js
@@ -1,9 +1,9 @@
 import React from 'react'
 import ReactECharts from 'echarts-for-react'
-import {formatDate} from '../utility/Timestamps'
+import {formatDate} from '../../utils/Timestamps'
 import {liquityClient, last7DayBlocks, splitQuery} from '../LiquityData'
 import dayjs from 'dayjs'
-import {abbreviateNumber} from '../utility/StringFormatter'
+import { abbreviateNumber } from '../../utils/NumberUtils'
 
 class StakingLUSD extends React.Component {
     state = {
